+++ toolium/config_parser.py
@@ -16,20 +16,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-try:
-    # Python 3
-    import configparser
-except ImportError:
-    # Python 2.7
-    import ConfigParser as configparser
-try:
-    # Python 2.7
-    from StringIO import StringIO
-except ImportError:
-    # Python 3
-    from io import StringIO
 import logging
 import os
+
+# Python 2 and 3 compatibility
+from six.moves import configparser, StringIO
 
 
 class ExtendedConfigParser(configparser.ConfigParser):