+++ toolium/test/test_driver_wrapper.py
@@ -22,10 +22,11 @@ import unittest
 
 import mock
 from ddt import ddt, data, unpack
+from nose.tools import assert_equal, assert_not_equal, assert_in, assert_raises
 
+from toolium.config_files import ConfigFiles
 from toolium.driver_wrapper import DriverWrapper
 from toolium.driver_wrappers_pool import DriverWrappersPool
-from toolium.config_files import ConfigFiles
 
 
 class DriverWrapperCommon(unittest.TestCase):
@@ -56,84 +57,84 @@ class DriverWrapperPropertiesTests(DriverWrapperCommon):
     def test_configure_properties(self):
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         self.wrapper.configure_properties()
-        self.assertEqual('firefox', self.wrapper.config.get('Browser', 'browser'))  # get last value
-        self.assertEqual('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
-        self.assertEqual(None, self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
+        assert_equal('firefox', self.wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
+        assert_equal(None, self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_android(self):
         os.environ["Config_prop_filenames"] = 'android-properties.cfg'
         self.wrapper.configure_properties()
-        self.assertEqual('android', self.wrapper.config.get('Browser', 'browser'))  # get last value
-        self.assertEqual(None, self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
-        self.assertEqual('http://invented_url/Demo.apk',
-                         self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
+        assert_equal('android', self.wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal(None, self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
+        assert_equal('http://invented_url/Demo.apk',
+                     self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_two_files(self):
         os.environ["Config_prop_filenames"] = 'properties.cfg;android-properties.cfg'
         self.wrapper.configure_properties()
-        self.assertEqual('android', self.wrapper.config.get('Browser', 'browser'))  # get last value
-        self.assertEqual('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
-        self.assertEqual('http://invented_url/Demo.apk',
-                         self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
+        assert_equal('android', self.wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
+        assert_equal('http://invented_url/Demo.apk',
+                     self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_two_files_android_first(self):
         os.environ["Config_prop_filenames"] = 'android-properties.cfg;properties.cfg'
         self.wrapper.configure_properties()
-        self.assertEqual('firefox', self.wrapper.config.get('Browser', 'browser'))  # get last value
-        self.assertEqual('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
-        self.assertEqual('http://invented_url/Demo.apk',
-                         self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
+        assert_equal('firefox', self.wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('5', self.wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
+        assert_equal('http://invented_url/Demo.apk',
+                     self.wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_system_property(self):
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         os.environ["Browser_browser"] = 'opera'
         self.wrapper.configure_properties()
-        self.assertEqual('opera', self.wrapper.config.get('Browser', 'browser'))
+        assert_equal('opera', self.wrapper.config.get('Browser', 'browser'))
 
     def test_configure_properties_file_default_file(self):
         self.wrapper.configure_properties()
-        self.assertEqual('firefox', self.wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.wrapper.config.get('Browser', 'browser'))
 
     def test_configure_properties_file_not_found(self):
         os.environ["Config_prop_filenames"] = 'notfound-properties.cfg'
-        with self.assertRaises(Exception) as cm:
+        with assert_raises(Exception) as cm:
             self.wrapper.configure_properties()
-        self.assertIn('Properties config file not found', str(cm.exception))
+        assert_in('Properties config file not found', str(cm.exception))
 
     def test_configure_properties_no_changes(self):
         # Configure properties
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         self.wrapper.configure_properties()
-        self.assertEqual('firefox', self.wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.wrapper.config.get('Browser', 'browser'))
 
         # Modify property
         new_browser = 'opera'
         self.wrapper.config.set('Browser', 'browser', new_browser)
-        self.assertEqual(new_browser, self.wrapper.config.get('Browser', 'browser'))
+        assert_equal(new_browser, self.wrapper.config.get('Browser', 'browser'))
 
         # Trying to configure again
         self.wrapper.configure_properties()
 
         # Configuration has not been initialized
-        self.assertEqual(new_browser, self.wrapper.config.get('Browser', 'browser'))
+        assert_equal(new_browser, self.wrapper.config.get('Browser', 'browser'))
 
     def test_configure_properties_change_configuration_file(self):
         # Configure properties
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         self.wrapper.configure_properties()
-        self.assertEqual('firefox', self.wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.wrapper.config.get('Browser', 'browser'))
 
         # Modify property
         new_browser = 'opera'
         self.wrapper.config.set('Browser', 'browser', new_browser)
-        self.assertEqual(new_browser, self.wrapper.config.get('Browser', 'browser'))
+        assert_equal(new_browser, self.wrapper.config.get('Browser', 'browser'))
 
         # Change file and try to configure again
         os.environ["Config_prop_filenames"] = 'android-properties.cfg'
         self.wrapper.configure_properties()
 
         # Check that configuration has been initialized
-        self.assertEqual('android', self.wrapper.config.get('Browser', 'browser'))
+        assert_equal('android', self.wrapper.config.get('Browser', 'browser'))
 
 
 class DriverWrapperLoggerTests(DriverWrapperCommon):
@@ -147,7 +148,7 @@ class DriverWrapperLoggerTests(DriverWrapperCommon):
     def test_configure_logger(self):
         os.environ["Config_log_filename"] = 'logging.conf'
         self.wrapper.configure_logger()
-        self.assertEqual('DEBUG', logging.getLevelName(self.wrapper.logger.getEffectiveLevel()))
+        assert_equal('DEBUG', logging.getLevelName(self.wrapper.logger.getEffectiveLevel()))
 
     def test_configure_logger_file_not_configured(self):
         self.wrapper.configure_logger()
@@ -163,36 +164,36 @@ class DriverWrapperLoggerTests(DriverWrapperCommon):
         os.environ["Config_log_filename"] = 'logging.conf'
         self.wrapper.configure_logger()
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
-        self.assertEqual('DEBUG', logging.getLevelName(logger.level))
+        assert_equal('DEBUG', logging.getLevelName(logger.level))
 
         # Modify property
         new_level = 'INFO'
         logger.setLevel(new_level)
-        self.assertEqual(new_level, logging.getLevelName(logger.level))
+        assert_equal(new_level, logging.getLevelName(logger.level))
 
         # Trying to configure again
         self.wrapper.configure_logger()
 
         # Configuration has not been initialized
-        self.assertEqual(new_level, logging.getLevelName(logger.level))
+        assert_equal(new_level, logging.getLevelName(logger.level))
 
     def test_configure_logger_change_configuration_file(self):
         # Configure logger
         os.environ["Config_log_filename"] = 'logging.conf'
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
-        self.assertEqual('DEBUG', logging.getLevelName(logger.level))
+        assert_equal('DEBUG', logging.getLevelName(logger.level))
 
         # Modify property
         new_level = 'INFO'
         logger.setLevel(new_level)
-        self.assertEqual(new_level, logging.getLevelName(logger.level))
+        assert_equal(new_level, logging.getLevelName(logger.level))
 
         # Change file and try to configure again
         os.environ["Config_log_filename"] = 'warn-logging.conf'
         self.wrapper.configure_logger()
 
         # Check that configuration has been initialized
-        self.assertEqual('WARNING', logging.getLevelName(logger.level))
+        assert_equal('WARNING', logging.getLevelName(logger.level))
 
 
 mobile_tests = (
@@ -251,13 +252,13 @@ class DriverWrapperTests(DriverWrapperCommon):
         new_wrapper.config.set('Browser', 'browser', new_browser)
 
         # Check that wrapper and new_wrapper are different
-        self.assertEqual(first_browser, self.driver_wrapper.config.get('Browser', 'browser'))
-        self.assertEqual(new_browser, new_wrapper.config.get('Browser', 'browser'))
-        self.assertNotEqual(self.driver_wrapper, new_wrapper)
+        assert_equal(first_browser, self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal(new_browser, new_wrapper.config.get('Browser', 'browser'))
+        assert_not_equal(self.driver_wrapper, new_wrapper)
 
     def test_configure_no_changes(self):
         # Check previous values
-        self.assertEqual('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
 
         # Modify wrapper
         self.driver_wrapper.config.set('Browser', 'browser', 'opera')
@@ -266,11 +267,11 @@ class DriverWrapperTests(DriverWrapperCommon):
         self.driver_wrapper.configure()
 
         # Configuration has not been initialized
-        self.assertEqual('opera', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('opera', self.driver_wrapper.config.get('Browser', 'browser'))
 
     def test_configure_change_configuration_file(self):
         # Check previous values
-        self.assertEqual('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
 
         # Modify wrapper
         self.driver_wrapper.config.set('Browser', 'browser', 'opera')
@@ -280,7 +281,7 @@ class DriverWrapperTests(DriverWrapperCommon):
         self.driver_wrapper.configure()
 
         # Check that configuration has been initialized
-        self.assertEqual('android', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('android', self.driver_wrapper.config.get('Browser', 'browser'))
 
     @mock.patch('toolium.driver_wrapper.ConfigDriver')
     def test_connect(self, ConfigDriver):
@@ -292,18 +293,18 @@ class DriverWrapperTests(DriverWrapperCommon):
 
         # Check the returned driver
         self.driver_wrapper.configure()
-        self.assertEqual(expected_driver, self.driver_wrapper.connect(maximize=False))
+        assert_equal(expected_driver, self.driver_wrapper.connect(maximize=False))
 
         # Check that the wrapper has been configured
-        self.assertEqual('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
-        self.assertEqual('DEBUG', logging.getLevelName(logger.level))
+        assert_equal('DEBUG', logging.getLevelName(logger.level))
 
     @data(*mobile_tests)
     @unpack
     def test_is_mobile_test(self, browser, is_mobile):
         self.driver_wrapper.config.set('Browser', 'browser', browser)
-        self.assertEqual(is_mobile, self.driver_wrapper.is_mobile_test())
+        assert_equal(is_mobile, self.driver_wrapper.is_mobile_test())
 
     @data(*web_tests)
     @unpack
@@ -313,10 +314,10 @@ class DriverWrapperTests(DriverWrapperCommon):
             self.driver_wrapper.config.set('AppiumCapabilities', 'app', appium_app)
         if appium_browser_name is not None:
             self.driver_wrapper.config.set('AppiumCapabilities', 'browserName', appium_browser_name)
-        self.assertEqual(is_web, self.driver_wrapper.is_web_test())
+        assert_equal(is_web, self.driver_wrapper.is_web_test())
 
     @data(*maximizable_browsers)
     @unpack
     def test_is_maximizable(self, browser, is_maximizable):
         self.driver_wrapper.config.set('Browser', 'browser', browser)
-        self.assertEqual(is_maximizable, self.driver_wrapper.is_maximizable())
+        assert_equal(is_maximizable, self.driver_wrapper.is_maximizable())