+++ toolium/test/test_utils.py
@@ -21,6 +21,7 @@ import unittest
 
 import mock
 from ddt import ddt, data, unpack
+from nose.tools import assert_is_none, assert_equal, assert_raises
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 
@@ -74,7 +75,7 @@ class UtilsTests(unittest.TestCase):
             self.driver_wrapper.config.set('AppiumCapabilities', 'app', appium_app)
         if appium_browser_name:
             self.driver_wrapper.config.set('AppiumCapabilities', 'browserName', appium_browser_name)
-        self.assertEqual(self.utils.get_safari_navigation_bar_height(), bar_height)
+        assert_equal(self.utils.get_safari_navigation_bar_height(), bar_height)
 
     def test_get_window_size_android_native(self):
         # Configure driver mock
@@ -83,7 +84,7 @@ class UtilsTests(unittest.TestCase):
         self.driver_wrapper.config.set('Browser', 'browser', 'android')
         self.driver_wrapper.config.set('AppiumCapabilities', 'app', 'C:/Demo.apk')
 
-        self.assertEqual(self.utils.get_window_size(), window_size)
+        assert_equal(self.utils.get_window_size(), window_size)
 
     def test_get_window_size_android_web(self):
         # Configure driver mock
@@ -93,7 +94,7 @@ class UtilsTests(unittest.TestCase):
         self.driver_wrapper.config.set('Browser', 'browser', 'android')
         self.driver_wrapper.config.set('AppiumCapabilities', 'browserName', 'chrome')
 
-        self.assertEqual(self.utils.get_window_size(), window_size)
+        assert_equal(self.utils.get_window_size(), window_size)
 
     def test_get_native_coords_android_web(self):
         # Configure driver mock
@@ -108,7 +109,7 @@ class UtilsTests(unittest.TestCase):
 
         web_coords = {'x': 105, 'y': 185}
         native_coords = {'x': 52.5, 'y': 92.5}
-        self.assertEqual(self.utils.get_native_coords(web_coords), native_coords)
+        assert_equal(self.utils.get_native_coords(web_coords), native_coords)
 
     def test_get_native_coords_ios_web(self):
         # Configure driver mock
@@ -120,7 +121,7 @@ class UtilsTests(unittest.TestCase):
 
         web_coords = {'x': 105, 'y': 185}
         native_coords = {'x': 52.5, 'y': 156.5}
-        self.assertEqual(self.utils.get_native_coords(web_coords), native_coords)
+        assert_equal(self.utils.get_native_coords(web_coords), native_coords)
 
     def test_swipe_android_native(self):
         # Configure driver mock
@@ -205,25 +206,25 @@ class UtilsTests(unittest.TestCase):
         # Create element mock
         element = get_mock_element(x=250, y=40, height=40, width=300)
 
-        with self.assertRaises(Exception) as cm:
+        with assert_raises(Exception) as cm:
             self.utils.swipe(element, 50, 100)
-        self.assertEqual(str(cm.exception), 'Swipe method is not implemented in Selenium')
+        assert_equal(str(cm.exception), 'Swipe method is not implemented in Selenium')
 
     def test_get_element_none(self):
         element = self.utils.get_element(None)
-        self.assertIsNone(element)
+        assert_is_none(element)
 
     def test_get_element_webelement(self):
         web_element = WebElement(None, 1)
         element = self.utils.get_element(web_element)
-        self.assertEqual(web_element, element)
+        assert_equal(web_element, element)
 
     def test_get_element_pageelement(self):
         page_element = mock.MagicMock()
         page_element.element.return_value = 'mock_element'
 
         element = self.utils.get_element(page_element)
-        self.assertEqual('mock_element', element)
+        assert_equal('mock_element', element)
         page_element.element.assert_called_with()
 
     def test_get_element_locator(self):
@@ -233,7 +234,7 @@ class UtilsTests(unittest.TestCase):
 
         # Get element and assert response
         element = self.utils.get_element(element_locator)
-        self.assertEqual('mock_element', element)
+        assert_equal('mock_element', element)
         self.driver_wrapper.driver.find_element.assert_called_with(*element_locator)