+++ toolium/test_cases.py
@@ -136,7 +136,7 @@ class SeleniumTestCase(BasicTestCase):
         if not self.reuse_driver:
             SeleniumTestCase.driver = None
 
-    def assertScreenshot(self, element_or_selector, filename, threshold=0, exclude_elements=[], driver_wrapper=None):
+    def assert_screenshot(self, element_or_selector, filename, threshold=0, exclude_elements=[], driver_wrapper=None):
         """Assert that a screenshot of an element is the same as a screenshot on disk, within a given threshold.
 
         :param element_or_selector: either a CSS/XPATH selector as a string or a WebElement object.
@@ -148,10 +148,10 @@ class SeleniumTestCase(BasicTestCase):
         :param driver_wrapper: driver wrapper instance
         """
         file_suffix = self.get_method_name()
-        VisualTest(driver_wrapper).assertScreenshot(element_or_selector, filename, file_suffix, threshold,
+        VisualTest(driver_wrapper).assert_screenshot(element_or_selector, filename, file_suffix, threshold,
                                                     exclude_elements)
 
-    def assertFullScreenshot(self, filename, threshold=0, exclude_elements=[], driver_wrapper=None):
+    def assert_full_screenshot(self, filename, threshold=0, exclude_elements=[], driver_wrapper=None):
         """Assert that a driver screenshot is the same as a screenshot on disk, within a given threshold.
 
         :param filename: the filename for the screenshot, which will be appended with ``.png``
@@ -161,7 +161,7 @@ class SeleniumTestCase(BasicTestCase):
         :param driver_wrapper: driver wrapper instance
         """
         file_suffix = self.get_method_name()
-        VisualTest(driver_wrapper).assertScreenshot(None, filename, file_suffix, threshold, exclude_elements)
+        VisualTest(driver_wrapper).assert_screenshot(None, filename, file_suffix, threshold, exclude_elements)
 
 
 class AppiumTestCase(SeleniumTestCase):